Fixed Floating error, and went back to floats since it's faster (Recommended by Ammar)

+++ includes/object.h
@@ -6,7 +6,7 @@
 /*   By: bismail <bismail@student.42.fr>            +#+  +:+       +#+        */
 /*                                                +#+#+#+#+#+   +#+           */
 /*   Created: 2025/02/10 15:04:22 by abueskander       #+#    #+#             */
-/*   Updated: 2025/02/12 10:45:34 by bismail          ###   ########.fr       */
+/*   Updated: 2025/02/12 11:54:18 by bismail          ###   ########.fr       */
 /*                                                                            */
 /* ************************************************************************** */
 
@@ -23,10 +23,10 @@
 
 typedef struct s_tuple
 {
-        double X;
-        double Y;
-        double Z;
-        double W;
+        float X;
+        float Y;
+        float Z;
+        float W;
 }       t_tuple;
 
 enum e_pov
@@ -34,16 +34,16 @@ enum e_pov
         VECTOR,
         POINT
 };
-int floatcmp(double a, double b);
+int floatcmp(float a, float b);
 int tuplecmp(t_tuple *a, t_tuple *b);
 t_tuple *tuplesub(t_tuple *a, t_tuple *b);
 t_tuple *tupleadd(t_tuple *a, t_tuple *b);
-t_tuple *vector(double x, double y, double z);
-t_tuple *point(double x, double y, double z);
-t_tuple *tupleSdiv(t_tuple *a, double scale);
-t_tuple *tupleSmult(t_tuple *a, double scale);
+t_tuple *vector(float x, float y, float z);
+t_tuple *point(float x, float y, float z);
+t_tuple *tupleSdiv(t_tuple *a, float scale);
+t_tuple *tupleSmult(t_tuple *a, float scale);
 t_tuple *tuplenegt(t_tuple *a);
-double tuplemagnitude(t_tuple *a);
+float tuplemagnitude(t_tuple *a);
 t_tuple *tuplenormalize(t_tuple *a);
 
 #endif
